Add files via upload
This contains a sloppy console output of most of the ticket data I want.  I still need to find a way to output multiple custom fields in a single line of text so that it can be easily nested into a SQL INSERT statement.  I believe I need to get away from the second for loop containing the custom field ID, or use multiple loops containing the other IDs I need to include.