Add files via upload
The current database and accompanying scripts are storing ticket data as intended.  Date-times are in UTC.  When querying, remember to use the CONVERT_TZ mysql function.
Next steps are to bring in comment data as prescribed in the existing schema, as well as add error logging to output to a text file, not just the console.